requirements.txt
anyio==3.7.1
asgiref==3.5.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.1.1
django-allauth==0.51.0
django-ckeditor==6.5.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-ninja==0.19.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.4.0
ecdsa==0.18.0
email-validator==2.0.0.post2
filelock==3.12.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
oauthlib==3.2.1
Pillow==9.2.0
platformdirs==3.9.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.24.0
